update `iced` rev and migrate

+++ data/src/config.rs
@@ -5,11 +5,11 @@ pub mod state;
 pub mod theme;
 pub mod timezone;
 
-pub const MIN_SCALE: f64 = 0.8;
-pub const MAX_SCALE: f64 = 1.5;
+pub const MIN_SCALE: f32 = 0.8;
+pub const MAX_SCALE: f32 = 1.5;
 
 #[derive(Debug, Clone, Copy, Deserialize, Serialize, PartialEq)]
-pub struct ScaleFactor(f64);
+pub struct ScaleFactor(f32);
 
 impl Default for ScaleFactor {
     fn default() -> Self {
@@ -17,13 +17,13 @@ impl Default for ScaleFactor {
     }
 }
 
-impl From<f64> for ScaleFactor {
-    fn from(value: f64) -> Self {
+impl From<f32> for ScaleFactor {
+    fn from(value: f32) -> Self {
         ScaleFactor(value.clamp(MIN_SCALE, MAX_SCALE))
     }
 }
 
-impl From<ScaleFactor> for f64 {
+impl From<ScaleFactor> for f32 {
     fn from(value: ScaleFactor) -> Self {
         value.0
     }
